feat(core): consolidar versionado 1.0.2 y shutdown seguro
Genera ClientVersion.h desde CMake y propaga los directorios de headers generados a todos los targets, dejando la documentación y el plan alineados con la release 1.0.2. Refuerza CPartFile, scheduler y throttler con contadores y condiciones para que las tareas asíncronas se apaguen limpiamente y actualiza SearchList/EC para IDs seguros.\n\nTests: cmake --build . --config Debug && ctest --output-on-failure -C Debug

+++ src/TextClient.h
@@ -28,6 +28,7 @@
 
 #include "ExternalConnector.h"
 
+#include <cstdint>
 #include <map>
 
 
@@ -37,10 +38,10 @@ class CEC_SearchFile_Tag;
 class SearchFile {
 	public:
 		wxString sFileName;
-		unsigned long lFileSize;
+		uint64_t lFileSize;
 		CMD4Hash  nHash;
 		wxString  sHash;
-		long lSourceCount;
+		uint32_t lSourceCount;
 		bool bPresent;
 
 		SearchFile(const CEC_SearchFile_Tag *);
